test(tor, external-links): move tests to their respective packages
